added mongo and rabbit relationships

+++ hooks/hooks.py
@@ -6,14 +6,10 @@ import os
 import utils
 import ceilometer_utils
 
-config = config_get()
-
-service = "ceilometer"
-
 def install():
     utils.configure_source()
     packages = ['python-ceilometer', 'ceilometer-common', 'ceilometer-agent-central', 'ceilometer-collector', 'ceilometer-api']
-    utils.install(packages)
+    utils.install(*packages)
 
 def amqp_joined():
     utils.relation_set(username=ceilometer_utils.RABBIT_USER, vhost=ceilometer_utils.RABBIT_VHOST)
@@ -33,21 +29,45 @@ def get_rabbit_conf():
                 return conf
     return None
 
+def get_db_conf():
+    for relid in utils.relation_ids('shared-db'):
+        for unit in utils.relation_list(relid):
+            conf = {
+                "db_host": utils.relation_get('hostname', unit, relid),
+                "db_port": utils.relation_get('port', unit, relid),
+                "db_name": ceilometer_utils.CEILOMETER_DB
+                }
+            if None not in conf.itervalues():
+                return conf
+    return None
+
 def render_ceilometer_conf():
     context = get_rabbit_conf()
-    context['metering_secret'] = ceilometer_utils.get_shared_secret()
+    contextdb = get_db_conf()
 
-    if (context and os.path.exists(ceilometer_utils.CEILOMETER_CONF)):
+    if (context and contextdb and os.path.exists(ceilometer_utils.CEILOMETER_CONF)):
+        context['metering_secret'] = ceilometer_utils.get_shared_secret()
+        context['db_connection'] = "mongodb://"+contextdb["db_host"]+":"+contextdb["db_port"]+"/"+contextdb["db_name"]
+
         with open(ceilometer_utils.CEILOMETER_CONF, "w") as conf:
             conf.write(utils.render_template(os.path.basename(ceilometer_utils.CEILOMETER_CONF), context))
 
 def amqp_changed():
     render_ceilometer_conf()
-    utils.restart(['ceilometer-agent-central', 'ceilometer-collector'])
+    utils.restart(*ceilometer_utils.CEILOMETER_SERVICES)
+
+def db_joined():
+    utils.relation_set(ceilometer_database=ceilometer_utils.CEILOMETER_DB)
+
+def db_changed():
+    render_ceilometer_conf()
+    utils.restart(*ceilometer_utils.CEILOMETER_SERVICES)
 
 utils.do_hooks({
     "install": install,
     "amqp-relation-joined": amqp_joined,
-    "amqp-relation-changed": amqp_changed
+    "amqp-relation-changed": amqp_changed,
+    "shared-db-relation-joined": db_joined,
+    "shared-db-relation-changed": db_changed
 })
 sys.exit(0)
